Renaming test-dir to test-root
