+++ packages/test-helpers/src/index.ts
@@ -1,6 +1,6 @@
 export { stdout } from './stdout';
 export { createTmpDir } from './tmp-dir';
-export { testRoot } from './test-dir';
+export { testRoot } from './test-root';
 export { default as CheckupProject } from './checkup-fixturify-project';
 export { default as CheckupPluginProject } from './checkup-plugin-project';
 export { default as EmberProject } from './ember-cli-fixturify-project';
